refactor(store): split optimization-pass orchestration into RecordingStore.Optimization.cs
